feat: improve agent performance and security
- Remove CAP_NET_ADMIN from systemd service for better security
- Add securityContext restrictions in k8s daemonset
- Refactor TSDB to use background pruning task and JSONL storage
- Optimize collector metrics handling and error recording
- Add package build workflows for deb and rpm
- Improve console UI with better power and thermal reporting

+++ crates/agent-core/src/collectors/disk.rs
@@ -93,23 +93,23 @@ impl Collector for DiskCollector {
                 metrics
                     .disk_read_ops_total
                     .with_label_values(&[dev.as_str()])
-                    .inc_by(rd_ops_delta as u64);
+                    .inc_by(rd_ops_delta);
                 metrics
                     .disk_write_ops_total
                     .with_label_values(&[dev.as_str()])
-                    .inc_by(wr_ops_delta as u64);
+                    .inc_by(wr_ops_delta);
                 metrics
                     .disk_read_bytes_total
                     .with_label_values(&[dev.as_str()])
-                    .inc_by(rd_bytes_delta as u64);
+                    .inc_by(rd_bytes_delta);
                 metrics
                     .disk_written_bytes_total
                     .with_label_values(&[dev.as_str()])
-                    .inc_by(wr_bytes_delta as u64);
+                    .inc_by(wr_bytes_delta);
                 metrics
                     .disk_io_time_ms_total
                     .with_label_values(&[dev.as_str()])
-                    .inc_by(io_ms_delta as u64);
+                    .inc_by(io_ms_delta);
 
                 if dt > 0.0
                     && root_io_ms_delta.is_none()
